REFACTOR: Replaced Headless UI modal with Shadcn UI for skills modal
