+++ package.json
@@ -6,6 +6,7 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
+    "build-start": "next build && next start",
     "lint": "next lint",
     "docker:compose": "docker-compose -f docker/docker-compose.yml up --build"
   },
